Move inner classes out of Session to top level
