+++ sql/src/main/java/io/crate/protocols/postgres/PostgresWireProtocol.java
@@ -23,6 +23,7 @@
 package io.crate.protocols.postgres;
 
 import com.google.common.annotations.VisibleForTesting;
+import io.crate.action.sql.DescribeResult;
 import io.crate.action.sql.ResultReceiver;
 import io.crate.action.sql.SQLOperations;
 import io.crate.action.sql.Session;
@@ -579,7 +580,7 @@ class PostgresWireProtocol {
     private void handleDescribeMessage(ByteBuf buffer, Channel channel) {
         byte type = buffer.readByte();
         String portalOrStatement = readCString(buffer);
-        Session.DescribeResult describeResult = session.describe((char) type, portalOrStatement);
+        DescribeResult describeResult = session.describe((char) type, portalOrStatement);
         Collection<Field> fields = describeResult.getFields();
         DataType[] parameterTypes = describeResult.getParameters();
         if (parameterTypes != null) {
@@ -687,7 +688,7 @@ class PostgresWireProtocol {
         try {
             session.parse("", query, Collections.emptyList());
             session.bind("", "", Collections.emptyList(), null);
-            Session.DescribeResult describeResult = session.describe('P', "");
+            DescribeResult describeResult = session.describe('P', "");
             List<Field> fields = describeResult.getFields();
 
             Function<Throwable, Exception> wrapError = SQLExceptions.forWireTransmission(
